# fix: schema loading performance