### What does this PR do?
reduce the time to resolve schemas content
replace `Array` with `Set`
performance of checking if the `Set.has` item is better than `Array.indexOf`

### What issues does this PR fix or reference?
no ref

### Is it tested? How?
manually with +-20 large schemas (connected between each other)
loading time from 3.5sec decreased to 1.5sec
`yarn test` is one second (20sec->19sec) faster on my side - but it's not really a reliable claim :)